+++ R/aniframe_kin_is.R
@@ -1,15 +1,15 @@
 #' Check if object is an aniframe_kin
 #'
 #' @param x An object to test
-#' @return Logical: TRUE if x inherits from aniframe
+#' @return Logical: TRUE if x inherits from aniframe_kin
 #' @examples
 #' kin <- calculate_kinematics(
-#'   anicore::example_aniframe(n_obs = 20, n_individuals = 1, n_keypoints = 1)
+#'   anicore::example_anipoint(n_obs = 20, n_individuals = 1, n_keypoints = 1)
 #' )
 #' is_aniframe_kin(kin)
 #'
-#' # An aniframe without kinematics is not one
-#' is_aniframe_kin(anicore::example_aniframe(n_obs = 3))
+#' # An anipoint without kinematics is not one
+#' is_aniframe_kin(anicore::example_anipoint(n_obs = 3))
 #' @export
 is_aniframe_kin <- function(x) {
   inherits(x, "aniframe_kin")